+++ tests/test_streamer.py
@@ -1,7 +1,7 @@
 from datastreams.streamer import Streamer
 from subgrounds.subgrounds import Subgrounds
 
-# Run a sample Streamer
+# Run Streamer over a list of subgraph IDs
 hosted_query_ids = [
                 'https://api.thegraph.com/subgraphs/name/cowprotocol/cow',
                 'https://api.thegraph.com/subgraphs/name/cowprotocol/cow-gc'
@@ -17,3 +17,6 @@ for endpoint in hosted_query_ids:
     dataStreamer = Streamer(sub, endpoint, subgraph)
     dataStreamer.runStreamer()
     sub_firehose_data.append(dataStreamer)
+
+
+# NOTE THIS HAS BEEN INTEGRATED INTO test_datastream.py test. Can safely delete this file.