+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ Changelog
 Version 0.1
 ===========
 
-- Feature A added
-- FIX: nasty bug #1729 fixed
-- add your changes here!
+- Added test_streamer.py to test the streamer class functionality
+- Added DataStream.py class as an execution layer interface to streamer
+- Added test_datastream.py to test the DataStream class functionality
+- Added documentation via Sphinx & tox